Moved eslint packages to dev dependencies
NPM's (Node Package Manager) dependency tree was flagging some eslint
plugin packages as invalid. This was because the plugin packages
(such as eslint-plugin-react) were listed as regular dependencies while
eslint was listed as a dev dependency. Eslint was fully functional
during this time, but the dependency tree did not know that. It has
now been fixed.