Factor out data EntryCell components from line items
This allows the developer to reuse EntryCell functionality in both the
AdditionLineItem components (used to insert new debts in the Main.jsx
component's Debts state) and the MorphingLineItem components (used to
modify existing debts in the Main.jsx component's Debt state).

This change also allows the developer to check for valid inputs, and
build consistent styling in a single location.