Create handleDeleteDebt functionality
handleDeleteDebt is one of two closures returned from
Main.jsx component's deleteHandler method. (The other function
object will be defined in a later commit, and involves the Edit
icon. It will turn the DisplayLineItem into a
MorphingLineItem.)

Closures returned from deleteHandler are bound to an index
within Main.jsx component's Debts state. They are bound when the
DebtTable component renders any of the DisplayLineItem components.

handleDeleteDebt is invoked whenever a user clicks on the Delete
icon within a DisplayLineItem component.