Rewrite line item rendering to delegate to header state
Inside Main.jsx comopnent, rename debtsMetadata to headerInfo. This
name is cleaner and easier to understand.

Change morphingHandler to accept a second argument {field}. Correspond
this to the field names present in each line item. This allows
morphingHandler to bind a field name to the handleEditChanges closure
during DebtTable.render, instead of assuming that a data-field
attribute will exist on the event.target.

Inside DebtTable's getMorphingLine() and getDisplayLine() methods,
delete the old code that manually specifies which field names to
render elements for. Replace it with an Array.prototype.map() method
that renders an element for each field name within headerInfo. This
ensures that morphing components and display components show the
same field names, and couples DebtTable's rendering logic to Main.jsx
state, which prevents DebtTable from rendering fields that don't exist
or skipping fields.