Patch addDebtContainer methods
A syntax error caused handleNewDebt and handleClearState to be passed
as children to AdditionBar. This change fixes the syntax error so
that the two methods are passed as properties instead.

Another syntax error caused handleNewDebt and handleClearState to
be passed as strings to AdditionBar's onClick method. This change
fixes the syntax error so that the two methods are passed as functions
instead.