Add header row to DebtTable
Add displayString property to Main.jsx headerInfo state. This
allows DebtTable to render a user-facing field name (printing
out 'Interest Rate' as opposed to 'interestRate').

Create getHeaderLine() method in DataTable. Render a div
for each field name in the headerInfo state. Style it like
the other line items, but add a bold font weight, and grey
background to distinguish it from the others.