Add payment schedule calculator
Optionally allow user to specify an accelerator amount (what they
would like to pay on top of their minimum monthly payments) and a
payment method (snowball means apply accelerator to the smallest
balance loan first and avalanche means apply to highest interest
first).

Set accelerator default to 0 and payment method default to snowball.

Accept an array of debt objects and output that same array but
with a schedule property defined on each debt. Schedule property
is an array of monthly payments objects, each of which specifies
the total monthly payment, the payment towards interest, the
payment towards principal reduction, and the amount leftover at the
end.