Patch visual bug for chartIndividualDebtsOverTime()
Previously chartIndividualDebtsOverTime() would visualize debts
in order of priority. This could mean that a bigger debt could
get visualized on top of a smaller debt, blocking the tooltip
for the smaller debt from rendering and also messing up the
color scheme.

This change orders the debts from largest balance to smallest
balance before visualizing the individual debts. This allows
smaller debts to get rendered on top of larger debts.