Add dir="ltr" to html tag; needed for Fabric UI
Fabric UI's Slider and ChoiceGroup components require an explicit
direction property to be specified on the global html tag in
order to work.

This is likely due to a conditional CSS properties that only
work when the html[dir="ltr"] evaluates as true. For example:

html[dir="ltr"] ms-Slider-line span {
    float: left;
}

This change allows Fabric UI components to work as expected.