Set both d3 chart axes to a constant range
Add getAxisScales() method to calculate scales for both axes
whenever Graph component updates. Set yScale based on a debt
parameter and a height parameter, and set xScale based on a
months parameter and a width parameter.

Before chart axes would recalculate whenever the accelerator
changed. Not only was this unsightly, it also made it difficult
for the user to determine how much quicker they could pay off
their debts with higher accelerator amounts. If the user specifies
an accelerator amount of 0, then the area would take up the
entire horizontal space. If the user specifies an accelerator
amount of 1000, then the area would still up the entire
horizontal space. This is true if the debt is paid off in 30
months with no accelerator, and if the debt is paid off in 20
months with an accelerator of 1000.

Now the x-axis and y-axis will be fixed. This means that a debt
being paid off in 20 months will take up less area and will look
smaller than a debt paid off in 30 months.

The y-axis will look exactly the same. The y-axis will be
fixed at total starting balance for visualizing total debt, and
at the highest debt in month 0 for visualizing individual
balances.

The x-axis will act differently. Now its range will be go from
0 months to the time it takes to pay off debts with no
accelerator. If it takes 30 months to pay off all debts with
no accelerator, then the x-axis will range from 0 to 30.